requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
google-cloud-storage==2.9.0
cloud-sql-python-connector==1.2.4
sqlalchemy==2.0.17
pymysql==1.0.3
Flask-sqlalchemy==3.0.5
flask-login==0.6.2
google-cloud-storage==2.9.0
plotly==5.15.0
pandas==1.5.3